+++ src/app/(with-header)/myprofile/_components/MyWIneKebabDropDown .tsx
@@ -5,8 +5,8 @@ import Image from 'next/image';
 import { toast } from 'react-toastify';
 import Dropdown from '@/components/Dropdown';
 import Modal from '@/components/modal/Modal';
-import PatchWineForm from '@/components/modal/PatchWineForm';
-import DeleteWineForm from '@/components/modal/DeleteWineModal';
+import PatchWineModal from '@/components/modal/PatchWineModal';
+import DeleteModal from '@/components/modal/DeleteModal';
 import { fetchDeleteWine } from '@/lib/fetchWines';
 import kebab from '@/assets/icons/menu.svg';
 
@@ -90,11 +90,11 @@ export default function MyWIneKebabDropDown({
         }`}
       >
         <div className='custom-scrollbar max-h-[90vh] overflow-y-auto'>
-          <PatchWineForm onClose={closeEditModal} id={`${id}`} wineInitialData={wineInitialData} editMyWine={editMyWine} />
+          <PatchWineModal onClose={closeEditModal} id={`${id}`} wineInitialData={wineInitialData} editMyWine={editMyWine} />
         </div>
       </Modal>
       <Modal isOpen={isDeleteModalOpen} setIsOpen={setIsDeleteModalOpen} className='rounded-2xl mobile:mx-auto mobile:h-[172px] mobile:max-w-[353px]'>
-        <DeleteWineForm onClose={closeDeleteModal} onDelete={handleDeleteWine} />
+        <DeleteModal onClose={closeDeleteModal} onDelete={handleDeleteWine} />
       </Modal>
     </div>
   );
